Move the Rate subclasses into their own modules
